Size the solver against actual cell count, not init points
Root cause of a second OOM on scene0140_00: "nonfrozen" densifies to 3x the init points, so that
scene has ~1.1M CELLS, not the 373k init points I had been sizing every buffer against. An fp32
(P, 7, 512) basis is then 16 GB and allocation fails immediately. Every memory estimate in this
work up to now was computed from the wrong number.

Fixes:
  - the augmented basis in prepare() is fp16 (build_blocks upcasts chunk-wise, so the einsum
    still runs in fp32);
  - pruning now happens BEFORE the basis is allocated, so the full edge set -- 383,374,495 edges
    on scene0140_00 -- is released rather than held alongside it.

Verified unchanged: the operator still matches the matrix-free reference at 5.7e-07.

Also records the real scale: edge count grows with cells TIMES views, and scene0140_00 at 215
views reaches 383M edges against scene0347_00's 16.2M at 54 views -- a 24x spread across scenes
in the same benchmark. Any per-scene budget has to be set from the measured count, not assumed.

+++ gram_blocks.py
@@ -229,8 +229,8 @@ def build_blocks(keys, svals, U, P, K, edge_chunk=200_000):
     B = torch.empty(E, K, K, device=U.device, dtype=torch.float32)
     for s in range(0, E, edge_chunk):
         e = min(s + edge_chunk, E)
-        Uj = U[j[s:e]]                       # (chunk, K, D)
-        Ul = U[l[s:e]]
+        Uj = U[j[s:e]].float()               # (chunk, K, D); U may be fp16 to fit
+        Ul = U[l[s:e]].float()
         B[s:e] = torch.einsum("ekd,eld->ekl", Uj, Ul) * svals[s:e, None, None]
         del Uj, Ul
     return BlockSparseHessian(j, l, B, P, K)
